refactor cost bigunt

+++ database/migrations/2024_01_01_090155_create_contract_details_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('harm_type_id'); 
             $table->foreign('harm_type_id')->references('id')->on('harm_types');
             
-            $table->integer('cost');
+            $table->bigInteger('cost');
             $table->integer('repeat')->nullable();
             $table->timestamps();
         });
